+++ routes/maxRatingRouter.js
@@ -25,13 +25,24 @@ maxRatingRouter.get("/:id/update", async(req, res) => {
 maxRatingRouter.post("/:id/update", async(req, res) => {
     const { id } = req.params;
     const newValue = req.body.updateTitle;
-    await db.deleteprogrammermaxratingbymaxratingId(id);
-    await db.updateRating(id, newValue);
-    res.redirect("/MaxRating");
+    const { password } = req.body;
+
+    if (password === process.env.password) {
+        await db.deleteprogrammermaxratingbymaxratingId(id);
+        await db.updateRating(id, newValue);
+        res.redirect("/MaxRating");
+    } else {
+        res.redirect("/MaxRating");
+    }
 });
 
 maxRatingRouter.post("/:id/delete", async(req, res) => {
     const { id } = req.params;
+    const { password } = req.body;
+
+    if (!password || password !== process.env.password) {
+        return res.status(403).send("Invalid Password");
+    } 
     try {
         await db.deleteprogrammermaxratingbymaxratingId(id);
         await db.deleteRatingAtid(id);
